+++ R/baidu.R
@@ -1,25 +1,38 @@
 ##' @rdname translate
 ##' @export
 baidu_translate <- function(x, from = 'en', to = 'zh') {
-    vapply(x, .baidu_translate, from = from, to = to, FUN.VALUE = character(1))
+    x <- gsub("\\s*\n+\\s*", " ", as.character(x), perl = use_perl())
+    text <- paste0(x, collapse="\n")
+    #vectorize_translator(x, 
+    #  .fun = .baidu_translate, 
+    #  from = from, to = to)
+    .translate(text, .fun = .baidu_translate, from = from, to = to)
 }
 
+
 ##' @importFrom jsonlite fromJSON
-##' @importFrom httr modify_url
+##' @importFrom httr2 request
+##' @importFrom httr2 req_url_query
+##' @importFrom httr2 req_perform
+##' @importFrom httr2 resp_body_json
 .baidu_translate <- function(x, from = 'en', to = 'zh') {
-    url <- httr::modify_url("http://api.fanyi.baidu.com/api/trans/vip/translate",
-            query = baidu_translate_query(x, from = from, to = to)
-        )
-    url <- url(url, encoding = "utf-8")
-    res <- jsonlite::fromJSON(url)
-    
-    return(res$trans_result$dst)
+    url_prefix <- "http://api.fanyi.baidu.com/api/trans/vip/translate"
+    query <- baidu_translate_query(x, from = from, to = to)
+    req <- httr2::request(url_prefix) |> httr2::req_url_query(!!!query) |>
+           httr2::req_perform()
+    res <- req |> httr2::resp_body_json()
+    structure(res, class = "baidu")
 }
+
+## @importFrom memoise memoise
+## .baidu_translate2 <- memoise(.baidu_translate)
+
+
 
 ##' @importFrom openssl md5
 baidu_translate_query <- function(x, from, to) {
     salt <- sample.int(1e+05, 1) 
-    .info <- get_translate_appkey()
+    .info <- get_translate_appkey('baidu')
     sign <- sprintf("%s%s%s%s", .info$appid, x, salt, .info$key)
     .sign <- openssl::md5(sign)
 
@@ -30,4 +43,13 @@ baidu_translate_query <- function(x, from, to) {
                 )
     return(.query)
 }
+
+
+##' @method get_translate_text baidu
+##' @export
+get_translate_text.baidu <- function(response) {
+    vapply(response$trans_result, function(x) x$dst, FUN.VALUE=character(1))
+}
+
+